refactor: App.jsx, Root.jsx 분리 및 상태관리 구현(테스트)
